cleaning fabric file and adding documentation

+++ fabfile.py
@@ -8,10 +8,21 @@ import urllib
 import markdown
 from distutils.filelist import FileList
 
-from fabric.api import env, execute, task, abort, warn, local, settings
+from fabric.api import env, execute, task, local, settings
+from fabric.api import warn as fabric_warn
+from fabric.api import abort as fabric_abort
 from fabric.context_managers import lcd, cd
 from fabric.colors import yellow, blue, red, green
 
+try:
+    import SimpleHTTPServer as srv
+except ImportError:
+    import http.server as srv
+
+try:
+    import SocketServer as socketserver
+except ImportError:
+    import socketserver
 
 SCRIPTS_JS = "js"
 STYLESHEETS_CSS = "css"
@@ -32,17 +43,6 @@ ABS_OUTPUT_PATH = os.path.join(ABS_ROOT_PATH, OUTPUT_DIR)
 
 SERVER_PORT = 8000
 SERVER_SUFFIXES = ['','.html','index.html']
-
-
-try:
-    import SimpleHTTPServer as srv
-except ImportError:
-    import http.server as srv
-
-try:
-    import SocketServer as socketserver
-except ImportError:
-    import socketserver
 
 
 class FabricHTTPRequestHandler(srv.SimpleHTTPRequestHandler):
@@ -69,12 +69,20 @@ class FabricHTTPRequestHandler(srv.SimpleHTTPRequestHandler):
         return path
 
 
+########################################################################################################################
+#
+# Tasks
+#
+########################################################################################################################
+
+
 @task
 def clean():
     """Clean up generated files."""
     if os.path.exists(ABS_OUTPUT_PATH):
-        __print_remove(OUTPUT_DIR)
-        shutil.rmtree(ABS_OUTPUT_PATH)
+        remove_tree(OUTPUT_DIR)
+    else:
+        notify(msg="{} directory not found - everything is already clean !".format(OUTPUT_DIR))
 
 
 @task
@@ -84,32 +92,28 @@ def generate():
     # clean first...
     execute(clean)
 
-    # generate medias, stylesheets, scripts
+    # process medias, stylesheets, scripts
     execute(process_medias)
     execute(process_assets)
     execute(process_stylesheets)
     execute(process_scripts)
 
-    # collect bower listed dependencies
-    execute(bower_install)
-
-    # generate pages
+    # process pages
     jinja_env = jinja2.Environment(loader=jinja2.FileSystemLoader(searchpath=TEMPLATES_DIR))
     jinja_context = {}
     jinja_pages = []
     for root, dirs, files in os.walk(TEMPLATES_DIR):
-        for dir in dirs:
-            if dir in PROTECTED_DIRS:
-                abort(red("you cannot use restricted names (i.e. {}) for naming your directories.".format(
-                    ','.join(PROTECTED_DIRS))))
-
-            if dir.startswith('_'):
+        for d in dirs:
+            if d in PROTECTED_DIRS:
+                abort("You cannot use restricted names (i.e. {}) for naming your directories.".format(
+                    ','.join(PROTECTED_DIRS)))
+            if d.startswith('_'):
                 jinja_context[os.path.relpath(root, TEMPLATES_DIR)] = {}
             else:
-                __print_create(os.path.join(OUTPUT_DIR, dir))
-                os.makedirs(os.path.join(ABS_OUTPUT_PATH, dir))
-        for file in files:
-            name, ext = os.path.splitext(file)
+                notify(action='create', path=os.path.join(OUTPUT_DIR, d))
+                os.makedirs(os.path.join(ABS_OUTPUT_PATH, d))
+        for f in files:
+            name, ext = os.path.splitext(f)
             basename = os.path.basename(root)
             relative_root = os.path.relpath(root, TEMPLATES_DIR)
             if name.startswith('_'):
@@ -117,7 +121,7 @@ def generate():
             elif basename.startswith('_'):
                 if ext == ".md":
                     data = markdown.Markdown(extensions=['meta'])
-                    with open(os.path.join(root, file), "r") as fh:
+                    with open(os.path.join(root, f), "r") as fh:
                         html = data.convert(fh.read())
                         meta = data.Meta
                         if not jinja_context[os.path.dirname(relative_root)].has_key(basename[1:]):
@@ -127,94 +131,25 @@ def generate():
                             'meta': meta
                         })
             elif ext == ".html":
-                jinja_pages.append(os.path.join(relative_root, file))
+                jinja_pages.append(os.path.join(relative_root, f))
             else:
-                __print_ignore(os.path.join(relative_root, file))
+                notify('ignore', src=os.path.join(relative_root, f))
 
     for page in jinja_pages:
         template = jinja_env.get_template(page)
         with open(os.path.join(OUTPUT_DIR, page), "wb") as fh:
-            __print_build(*os.path.split(page))
+            notify('build', src=page)
             breadcrumb = os.path.dirname(page).split(os.sep) if os.path.dirname(page) is not '.' else []
             fh.write(template.render(breadcrumb=breadcrumb, **jinja_context))
 
     # finally generate resources
-    execute(generate_resources)
+    execute(process_resources)
 
 
-def process_medias():
-    """Process medias"""
-
-    if os.path.exists(MEDIAS_DIR):
-        __print_copy(MEDIAS_DIR, os.path.join(OUTPUT_DIR, STATICS_DIR, MEDIAS_DIR))
-        shutil.copytree(MEDIAS_DIR, os.path.join(OUTPUT_DIR, STATICS_DIR, MEDIAS_DIR))
-
-
-def process_assets():
-    """Process assets"""
-
-    # check bower assets
-    execute(bower_install, force=False)
-
-    if not os.path.exists('assets.in'):
-        abort(red("Please specify an assets.in file."))
-
-    # load assets filter
-    filelist = FileList()
-    with open('assets.in', "r") as fh:
-        for line in fh:
-            filelist.process_template_line(line)
-
-    # copy assets
-    if os.path.exists(ASSETS_DIR):
-        for file in filelist.files:
-            output_dir = os.path.join(OUTPUT_DIR, os.path.dirname(file))
-            if not os.path.exists(output_dir):
-                __print_create(output_dir)
-                os.makedirs(output_dir)
-            __print_copy(file, os.path.join(OUTPUT_DIR, file))
-            shutil.copyfile(file, os.path.join(OUTPUT_DIR, file))
-
-
-def process_stylesheets():
-    """Generate stylesheets. Only CSS is exported."""
-
-    if os.path.exists(STYLESHEETS_DIR):
-        if os.path.exists(os.path.join(STYLESHEETS_DIR, STYLESHEETS_CSS)):
-            __print_copy(os.path.join(STYLESHEETS_DIR, STYLESHEETS_CSS), os.path.join(OUTPUT_DIR, STATICS_DIR, STYLESHEETS_CSS))
-            shutil.copytree(os.path.join(STYLESHEETS_DIR, STYLESHEETS_CSS), os.path.join(OUTPUT_DIR, STATICS_DIR, STYLESHEETS_CSS))
-        if os.path.exists(os.path.join(STYLESHEETS_DIR, STYLESHEETS_SCSS)):
-            local("scss --update {0}:{1} --load-path {2}".format(
-                os.path.join(STYLESHEETS_DIR, STYLESHEETS_SCSS),
-                os.path.join(OUTPUT_DIR, STATICS_DIR, STYLESHEETS_CSS),
-                ASSETS_DIR))
-
-
-def process_scripts():
-    """Generate scripts. Only JavaScript is exported"""
-
-    if os.path.exists(os.path.join(SCRIPTS_DIR, SCRIPTS_JS)):
-        __print_copy(os.path.join(SCRIPTS_DIR, SCRIPTS_JS), os.path.join(OUTPUT_DIR, STATICS_DIR, SCRIPTS_JS))
-        shutil.copytree(os.path.join(SCRIPTS_DIR, SCRIPTS_JS), os.path.join(OUTPUT_DIR, STATICS_DIR, SCRIPTS_JS))
-
-
-def generate_resources():
-    """Generate resources (directory are ignored)."""
-
-    for root, dirs, files in os.walk(RESOURCES_DIR):
-        for f in files:
-            if os.path.exists(os.path.join(OUTPUT_DIR, f)):
-                warn(red("File {} is in conflict - it will be skipped."))
-                continue
-            else:
-                __print_copy(os.path.join(root, f), os.path.join(OUTPUT_DIR, f))
-                shutil.copyfile(os.path.join(root, f), os.path.join(OUTPUT_DIR, f))
-
-
-@task()
-def bower_install(force=True):
-    """Install bower dependencies."""
-    if not os.path.exists(ASSETS_DIR) or force:
+@task
+def bower(action, force=True):
+    """Wrap bower tasks."""
+    if action == 'install' and (not os.path.exists(ASSETS_DIR) or force):
         local("bower install --config.directory={}".format(ASSETS_DIR))
 
 
@@ -240,15 +175,85 @@ def serve():
             execute(run_server)
 
 
+########################################################################################################################
+#
+# Sub tasks
+#
+########################################################################################################################
+
+
+def process_medias():
+    """Process medias"""
+
+    if os.path.exists(MEDIAS_DIR):
+        copy_tree(src=MEDIAS_DIR, dst=os.path.join(OUTPUT_DIR, STATICS_DIR, MEDIAS_DIR))
+    else:
+        warn("{} directory not found.".format(MEDIAS_DIR))
+
+
+def process_assets():
+    """Process assets"""
+
+    # check bower assets
+    if not os.path.exists(ASSETS_DIR):
+        abort("{} directory not found - please run > fab bower:install".format(ASSETS_DIR))
+
+    if not os.path.exists('assets.in'):
+        abort("Please specify an assets.in file.")
+
+    # load assets filter
+    filelist = FileList()
+    with open('assets.in', "r") as fh:
+        for line in fh:
+            filelist.process_template_line(line)
+
+    # copy assets
+    if os.path.exists(ASSETS_DIR):
+        for f in filelist.files:
+            copy_file(src=f, dst=os.path.join(OUTPUT_DIR, f), create_dir_if_not_exist=True)
+
+
+def process_stylesheets():
+    """Generate stylesheets. Only CSS is exported."""
+
+    if os.path.exists(STYLESHEETS_DIR):
+        if os.path.exists(os.path.join(STYLESHEETS_DIR, STYLESHEETS_CSS)):
+            copy_tree(src=os.path.join(STYLESHEETS_DIR, STYLESHEETS_CSS), dst=os.path.join(OUTPUT_DIR, STATICS_DIR, STYLESHEETS_CSS))
+        if os.path.exists(os.path.join(STYLESHEETS_DIR, STYLESHEETS_SCSS)):
+            local("scss --update {0}:{1} --load-path {2}".format(
+                os.path.join(STYLESHEETS_DIR, STYLESHEETS_SCSS),
+                os.path.join(OUTPUT_DIR, STATICS_DIR, STYLESHEETS_CSS),
+                ASSETS_DIR))
+
+
+def process_scripts():
+    """Generate scripts. Only JavaScript is exported"""
+
+    if os.path.exists(os.path.join(SCRIPTS_DIR, SCRIPTS_JS)):
+        copy_tree(src=os.path.join(SCRIPTS_DIR, SCRIPTS_JS), dst=os.path.join(OUTPUT_DIR, STATICS_DIR, SCRIPTS_JS))
+
+
+def process_resources():
+    """Generate resources (directory are ignored)."""
+
+    for root, dirs, files in os.walk(RESOURCES_DIR):
+        for f in files:
+            if os.path.exists(os.path.join(OUTPUT_DIR, f)):
+                warn("File {} is in conflict - it will be skipped.".format(f))
+                continue
+            else:
+                copy_file(src=os.path.join(root, f), dst=os.path.join(OUTPUT_DIR, f))
+
+
 def run_server():
     Handler = FabricHTTPRequestHandler
 
     try:
         httpd = socketserver.TCPServer(('', SERVER_PORT), Handler)
+        notify(action="serving", msg="at port %s" % SERVER_PORT)
     except OSError as e:
-        abort(red("Could not listen on post %s") % SERVER_PORT)
+        abort("Could not listen on post {}".format(SERVER_PORT))
 
-    print("Serving at port %s" % SERVER_PORT)
     try:
         httpd.serve_forever()
     except KeyboardInterrupt as e:
@@ -256,23 +261,51 @@ def run_server():
         httpd.socket.close()
 
 
-# --- stupid output functions ---
+########################################################################################################################
+#
+# Utilities
+#
+########################################################################################################################
 
 
-def __print_create(path):
-    print(yellow("creating ") + blue(os.path.join(path)))
+def copy_tree(src, dst):
+    notify('copy', src=src, dst=dst)
+    shutil.copytree(src, dst)
 
 
-def __print_copy(src, dst):
-    print(yellow("copying ") + blue(src) + " to " + blue(dst))
+def remove_tree(path):
+    notify('remove', path=path)
+    shutil.rmtree(path)
 
 
-def __print_remove(path):
-    print(yellow("removing ") + blue(path))
+def copy_file(src, dst, create_dir_if_not_exist=False):
+    if create_dir_if_not_exist and not os.path.exists(os.path.dirname(dst)):
+        output_dir = os.path.dirname(dst)
+        notify('create', path=output_dir)
+        os.makedirs(output_dir)
+    notify('copy', src=src, dst=dst)
+    shutil.copyfile(src, dst)
 
 
-def __print_build(root, file):
-    print(yellow("building ") + blue(os.path.join(root, file)))
+def warn(msg):
+    fabric_warn(red(msg))
 
-def __print_ignore(file):
-    print(green("ignoring ") + blue(file))
+
+def abort(msg):
+    fabric_abort(red(msg))
+
+def notify(action=None, **kwargs):
+    if action == "create":
+        print(yellow("creating ") + blue(kwargs['path']))
+    elif action == "copy":
+        print(yellow("copying ") + blue(kwargs['src']) + " to " + blue(kwargs['dst']))
+    elif action == 'build':
+        print(yellow("building ") + blue(kwargs['src']))
+    elif action == 'ignore':
+        print(green("ignoring ") + blue(kwargs['src']))
+    elif action == 'remove':
+        print(yellow("removing ") + blue(kwargs['path']))
+    elif action is not None:
+        print("{action} {msg}".format(action=action, msg=kwargs['msg']))
+    else:
+        print(kwargs['msg'])
